UX-437 Add portal injection, Convert overlays
