+++ libby/preview.js
@@ -1,0 +1,5 @@
+export default function() {
+  const portal = document.createElement('div');
+  portal.id = 'portal';
+  document.body.append(portal);
+}
